fix(db): fix all 40 Postgres functions to match actual schema
- Fix non-existent table references: anomalies → alerts, battery_cells → vehicle_live_state, charging_telemetry_readings → charge_telemetry_readings
- Fix column mismatches: energy_used_kwh → computed from rated range delta, avg_speed → speed_avg, energy_added_kwh → charge_energy_added, charge_rate_kw → charger_power, gas_equivalent_cost → computed estimate
- Fix drive_score/efficiency_score → computed from distance/range efficiency
- Fix ROUND(double, int) → ROUND(value::numeric, int) for PG compatibility
- Fix quoted reserved word 'time' in RETURNS TABLE signatures
- Fix Grafana route-efficiency and regen-efficiency dashboard queries
- All 40/40 functions pass runtime validation

+++ migrations/000071_add_fn_charging_hourly_distribution.up.sql
@@ -3,10 +3,10 @@ CREATE OR REPLACE FUNCTION fn_charging_hourly_distribution(
   p_from TIMESTAMPTZ DEFAULT NULL,
   p_to TIMESTAMPTZ DEFAULT NULL
 )
-RETURNS TABLE (hour DOUBLE PRECISION, sessions BIGINT) AS $$
+RETURNS TABLE (hour INTEGER, sessions BIGINT) AS $$
 BEGIN
   RETURN QUERY
-  SELECT EXTRACT(HOUR FROM cs.start_date), COUNT(*)::bigint
+  SELECT EXTRACT(HOUR FROM cs.start_date)::integer, COUNT(*)::bigint
   FROM charging_sessions cs
   WHERE cs.vehicle_id = p_vehicle_id
     AND (p_from IS NULL OR cs.start_date >= p_from)
